# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant Chain POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/modifiers:
    get:
      operationId: listModifierGroups
      tags: [menu]
      summary: List modifier groups for an outlet
      parameters:
        - name: outletId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Modifier groups with options
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModifierGroup"
    post:
      operationId: createModifierGroup
      tags: [menu]
      summary: Create a modifier group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifierGroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModifierGroup"
  /menu/modifiers/{groupId}:
    patch:
      operationId: updateModifierGroup
      tags: [menu]
      summary: Update a modifier group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifierGroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModifierGroup"
    delete:
      operationId: deleteModifierGroup
      tags: [menu]
      summary: Delete a modifier group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menu/modifiers/{groupId}/options:
    post:
      operationId: addModifierOption
      tags: [menu]
      summary: Add an option to a modifier group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifierOptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModifierOption"
  /menu/modifiers/{groupId}/options/{optionId}:
    patch:
      operationId: updateModifierOption
      tags: [menu]
      summary: Update a modifier option
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: optionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifierOptionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModifierOption"
    delete:
      operationId: deleteModifierOption
      tags: [menu]
      summary: Delete a modifier option
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: optionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Menu Item ↔ Modifier Group assignments ───────────────────────────────
  /menu/items/{itemId}/modifier-groups:
    get:
      operationId: listItemModifierGroups
      tags: [menu]
      summary: List modifier groups assigned to a menu item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Modifier groups with options
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModifierGroup"
    post:
      operationId: assignItemModifierGroup
      tags: [menu]
      summary: Assign a modifier group to a menu item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemModifierGroupInput"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemModifierGroupAssignment"
  /menu/items/{itemId}/modifier-groups/{groupId}:
    delete:
      operationId: unassignItemModifierGroup
      tags: [menu]
      summary: Remove a modifier group from a menu item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Areas ─────────────────────────────────────────────────────────────────
...
  /orders/{id}/items/{itemId}/modifiers:
    post:
      operationId: addOrderItemModifier
      tags: [orders]
      summary: Attach a modifier option to an order item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemModifierInput"
      responses:
        "201":
          description: Modifier attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItemModifier"

  # ── Kitchen Display ───────────────────────────────────────────────────────
...
components:
  schemas:
    ItemModifierGroupInput:
      type: object
      required: [modifierGroupId]
      properties:
        modifierGroupId:
          type: integer
    ItemModifierGroupAssignment:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        modifierGroupId:
          type: integer
...
    ModifierGroup:
      type: object
      required: [id, outletId, name, required, multiSelect, options]
      properties:
        id:
          type: integer
        outletId:
          type: integer
        name:
          type: string
        required:
          type: boolean
        multiSelect:
          type: boolean
        options:
          type: array
          items:
            $ref: "#/components/schemas/ModifierOption"
    ModifierGroupInput:
      type: object
      required: [outletId, name]
      properties:
        outletId:
          type: integer
        name:
          type: string
        required:
          type: boolean
          default: false
        multiSelect:
          type: boolean
          default: false
    ModifierGroupUpdate:
      type: object
      properties:
        name:
          type: string
        required:
          type: boolean
        multiSelect:
          type: boolean
    ModifierOption:
      type: object
      required: [id, groupId, name, priceAdjustment]
      properties:
        id:
          type: integer
        groupId:
          type: integer
        name:
          type: string
        priceAdjustment:
          type: number
    ModifierOptionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        priceAdjustment:
          type: number
          default: 0
    OrderItemModifier:
      type: object
      required: [id, orderItemId, modifierOptionId, name, priceAdjustment]
      properties:
        id:
          type: integer
        orderItemId:
          type: integer
        modifierOptionId:
          type: integer
        name:
          type: string
        priceAdjustment:
          type: number
    OrderItemModifierInput:
      type: object
      required: [modifierOptionId]
      properties:
        modifierOptionId:
          type: integer

    # ── Area schemas ──────────────────────────────────────────────────────
...